// src/shared/types.ts
interface ApiResultItem {
  type: "text" | "file" | "quote"
  contentText?: string
  contentBuffer?: ArrayBuffer
  name?: string
  mimeType?: string
}